extract TestConfig to helper package
